refactor: split rules in multiple files
